Update API base URL for production, integrate app expiry overlay, enhance profile completeness checks, and improve transaction handling in earnings screen
- Changed PROD_API_BASE_URL in .env to point to the live API.
- Added AppExpiryOverlay widget to block access after a set expiry date.
- Implemented profile completeness checks in various screens to guide users in completing their profiles.
- Enhanced the SellerEarningScreen to include a RefreshIndicator for transaction fetching.
- Updated SellerPaymentMethodFormWidget to refresh profile details after adding a payment method.
- Improved SellerProfileManagementScreen to show toast messages upon successful updates and check profile completeness.
- Modified SellerProfileVerificationScreen to handle profile completeness after submission.
- Added confirmation dialogs for deleting work items in SellerPortfolioScreen and SellerWorkScreen.
- Enhanced SellerWorksCubit to show toast messages for delete actions.

+++ lib/features/account/seller/work/views/seller_portfolio_screen.dart
@@ -8,6 +8,8 @@ import 'package:snip_fair/core/di/injector.dart';
 import 'package:snip_fair/core/presentation/theme/app_colors.dart';
 import 'package:snip_fair/core/presentation/widgets/app_text.dart';
 import 'package:snip_fair/core/presentation/widgets/custom_appbar.dart';
+import 'package:snip_fair/core/presentation/widgets/dialogs.dart';
+import 'package:snip_fair/core/utils/app_helper.dart';
 import 'package:snip_fair/features/account/seller/work/cubit/seller_works_cubit.dart';
 
 import 'package:snip_fair/features/account/seller/work/views/seller_work_form_view.dart';
@@ -157,9 +159,19 @@ class SellerPortfolioScreen extends StatelessWidget
                       return SellerWorkItemWidget(
                         work: work,
                         onDelete: () {
-                          context
-                              .read<SellerWorksCubit>()
-                              .deleteWorkItem(work.id!);
+                          AppHelper.showAppDialog(
+                            context,
+                            OnConfirmDialog(
+                              title: 'Delete Work ',
+                              content:
+                                  'Are you sure you want to delete this work item?',
+                              onConfirmed: (_) {
+                                context
+                                    .read<SellerWorksCubit>()
+                                    .deleteWorkItem(work.id!);
+                              },
+                            ),
+                          );
                         },
                         onEdit: () => SellerWorkFormWidget.show(context, work),
                       );
